fix SDL for static building

+++ scripts/make/SDL.sh
@@ -2,8 +2,12 @@
 set -e
 
 make -j4
-cp ./build/.libs/SDL2.dll ../000res
 cp ./build/.libs/libSDL2main.a ../000res
+if [ -f ./build/.libs/SDL2.dll ] ; then
+	cp ./build/.libs/SDL2.dll ../000res
+else
+	cp ./build/.libs/libSDL2.a ../000res
+fi
 
 # remove -lmingw32 and -Dmain=SDL_main from sdl2-config
 sed -e "s/-lmingw32//" sdl2-config > sdl2-config.new
